requirements.txt
asgiref==3.7.1
cloudinary==1.33.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-allauth==0.54.0
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==2.0
django-resized==1.0.2
django-richtextfield==1.6.1
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pydantic==1.10.8
PyJWT==2.7.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
urllib3==1.26.16
